refactor: overlay types and fix zIndex cannot work
