+++ src/packages/overlay/overlay.taro.tsx
@@ -1,32 +1,23 @@
 import React, { useState, FunctionComponent, useEffect } from 'react'
 import { CSSTransition } from 'react-transition-group'
-import { EnterHandler, ExitHandler } from 'react-transition-group/Transition'
 import classNames from 'classnames'
 import { View, ITouchEvent } from '@tarojs/components'
-import { BasicComponent, ComponentDefaults } from '@/utils/typings'
+import { ComponentDefaults } from '@/utils/typings'
 import { useLockScrollTaro } from '@/utils/use-lock-scoll-taro'
 import { harmony } from '@/utils/platform-taro'
+import { OverlayProps } from './types'
 
-export interface OverlayProps extends BasicComponent {
-  zIndex: number
-  duration: number
-  closeOnOverlayClick: boolean
-  visible: boolean
-  lockScroll: boolean
-  onClick: (event: ITouchEvent) => void
-  afterShow: () => void
-  afterClose: () => void
-}
-
-export const defaultOverlayProps = {
+export const defaultOverlayProps: OverlayProps = {
   ...ComponentDefaults,
   zIndex: 1000,
   duration: 300,
   closeOnOverlayClick: true,
   visible: false,
   lockScroll: true,
-  onClick: (event: ITouchEvent) => {},
-} as OverlayProps
+  onClick: () => {},
+  afterShow: () => {},
+  afterClose: () => {},
+}
 export const Overlay: FunctionComponent<
   Partial<OverlayProps> & Omit<React.HTMLAttributes<HTMLDivElement>, 'onClick'>
 > = (props) => {
@@ -43,28 +34,20 @@ export const Overlay: FunctionComponent<
     afterClose,
     onClick,
     ...rest
-  } = {
-    ...defaultOverlayProps,
-    ...props,
-  }
-  const classPrefix = `nut-overlay`
+  } = { ...defaultOverlayProps, ...props }
 
+  const classPrefix = 'nut-overlay'
   const [innerVisible, setInnerVisible] = useState(visible)
-
   const nodeRef = useLockScrollTaro(!!lockScroll && innerVisible)
 
   useEffect(() => {
-    if (visible) {
-      setInnerVisible(true)
-    } else {
-      setInnerVisible(false)
-    }
+    setInnerVisible(visible)
   }, [visible])
 
   const classes = classNames(classPrefix, className)
-
   const styles = {
     ...style,
+    zIndex,
   }
 
   const handleClick = (e: ITouchEvent) => {
@@ -73,32 +56,18 @@ export const Overlay: FunctionComponent<
     }
   }
 
-  const onHandleOpened: EnterHandler<HTMLElement | undefined> | undefined = (
-    e: HTMLElement
-  ) => {
-    afterShow && afterShow()
-  }
-
-  const onHandleClosed: ExitHandler<HTMLElement | undefined> | undefined = (
-    e: HTMLElement
-  ) => {
-    afterClose && afterClose()
-  }
-
-  function renderHarmony() {
-    return innerVisible ? (
-      <View
-        ref={nodeRef}
-        className={classes}
-        style={styles}
-        {...(rest as any)}
-        catchMove={lockScroll}
-        onClick={handleClick}
-      >
-        {children}
-      </View>
-    ) : null
-  }
+  const renderOverlay = () => (
+    <View
+      ref={nodeRef}
+      className={classes}
+      style={styles}
+      {...(rest as any)}
+      catchMove={lockScroll}
+      onClick={handleClick}
+    >
+      {children}
+    </View>
+  )
 
   return (
     <>
@@ -109,22 +78,13 @@ export const Overlay: FunctionComponent<
           unmountOnExit
           timeout={duration}
           in={innerVisible}
-          onEntered={onHandleOpened}
-          onExited={onHandleClosed}
+          onEntered={afterShow}
+          onExited={afterClose}
         >
-          <View
-            ref={nodeRef}
-            className={classes}
-            style={styles}
-            {...(rest as any)}
-            catchMove={lockScroll}
-            onClick={handleClick}
-          >
-            {children}
-          </View>
+          {renderOverlay()}
         </CSSTransition>
       ) : (
-        renderHarmony()
+        innerVisible && renderOverlay()
       )}
     </>
   )
