+++ src/packages/overlay/overlay.tsx
@@ -1,37 +1,27 @@
 import React, {
   FunctionComponent,
   MouseEvent,
-  MouseEventHandler,
   useEffect,
   useRef,
   useState,
 } from 'react'
 import { CSSTransition } from 'react-transition-group'
-import { EnterHandler, ExitHandler } from 'react-transition-group/Transition'
 import classNames from 'classnames'
-import { BasicComponent, ComponentDefaults } from '@/utils/typings'
+import { ComponentDefaults } from '@/utils/typings'
 import { useLockScroll } from '@/utils/use-lock-scroll'
+import { OverlayProps } from './types'
 
-export interface OverlayProps extends BasicComponent {
-  zIndex: number
-  duration: number
-  closeOnOverlayClick: boolean
-  visible: boolean
-  lockScroll: boolean | 'strict'
-  onClick: (event: MouseEvent) => void
-  afterShow: () => void
-  afterClose: () => void
-}
-
-export const defaultOverlayProps = {
+export const defaultOverlayProps: OverlayProps = {
   ...ComponentDefaults,
   zIndex: 1000,
   duration: 300,
   closeOnOverlayClick: true,
   visible: false,
   lockScroll: true,
-  onClick: (event: MouseEvent) => {},
-} as OverlayProps
+  onClick: () => {},
+  afterShow: () => {},
+  afterClose: () => {},
+}
 export const Overlay: FunctionComponent<
   Partial<OverlayProps> & React.HTMLAttributes<HTMLDivElement>
 > = (props) => {
@@ -48,50 +38,28 @@ export const Overlay: FunctionComponent<
     afterClose,
     onClick,
     ...rest
-  } = {
-    ...defaultOverlayProps,
-    ...props,
-  }
-  const classPrefix = `nut-overlay`
+  } = { ...defaultOverlayProps, ...props }
 
+  const classPrefix = 'nut-overlay'
   const [innerVisible, setInnerVisible] = useState(visible)
-
   const nodeRef = useRef(null)
 
   useEffect(() => {
-    if (visible) {
-      setInnerVisible(true)
-    } else {
-      setInnerVisible(false)
-    }
+    setInnerVisible(visible)
   }, [visible])
 
   const shouldLockScroll = !innerVisible ? false : lockScroll
-
   useLockScroll(nodeRef, shouldLockScroll)
-
   const classes = classNames(classPrefix, className)
-
   const styles = {
     ...style,
+    zIndex,
   }
 
-  const handleClick: MouseEventHandler<HTMLDivElement> = (e: MouseEvent) => {
+  const handleClick = (e: MouseEvent) => {
     if (closeOnOverlayClick) {
       onClick && onClick(e)
     }
-  }
-
-  const onHandleOpened: EnterHandler<HTMLElement | undefined> | undefined = (
-    e: HTMLElement
-  ) => {
-    afterShow && afterShow()
-  }
-
-  const onHandleClosed: ExitHandler<HTMLElement | undefined> | undefined = (
-    e: HTMLElement
-  ) => {
-    afterClose && afterClose()
   }
 
   return (
@@ -101,8 +69,8 @@ export const Overlay: FunctionComponent<
       unmountOnExit
       timeout={duration}
       in={innerVisible}
-      onEntered={onHandleOpened}
-      onExited={onHandleClosed}
+      onEntered={afterShow}
+      onExited={afterClose}
     >
       <div
         ref={nodeRef}
